Move utility out of the atomspace directory
