+++ opencog/atomutils/CMakeLists.txt
@@ -26,5 +26,6 @@ INSTALL (FILES
     ForeachChaseLink.h
     Foreach.h
     ForeachTwo.h
+    HandleMap.h
     DESTINATION "include/${PROJECT_NAME}/atomutils"
 )
