+++ opencog/atomutils/HandleMap.h
