+++ opencog/persist/file/SavingLoading.cc
@@ -33,7 +33,7 @@
 
 #include <opencog/atomspace/ClassServer.h>
 #include <opencog/atomspace/CountTruthValue.h>
-#include <opencog/atomspace/HandleMap.h>
+#include <opencog/atomutils/HandleMap.h>
 #include <opencog/atomspace/IndefiniteTruthValue.h>
 #include <opencog/atomspace/Link.h>
 #include <opencog/atomspace/Node.h>
