+++ opencog/spacetime/HandleToTemporalEntryMap.cc
@@ -25,7 +25,7 @@
 #include "HandleToTemporalEntryMap.h"
 
 #include <opencog/atomspace/Atom.h>
-#include <opencog/atomspace/HandleMap.h>
+#include <opencog/atomutils/HandleMap.h>
 
 using namespace opencog;
